requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.3.4
beautifulsoup4==4.7.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
class-registry==2.1.2
cryptography==2.4.1
cycler==0.10.0
Dajax==1.3
decorator==4.4.0
Django==2.1.7
django-admin-export==2.0
django-ajax-selects==1.7.1
django-autocomplete-light==3.3.0
django-bower==5.2.0
django-braces==1.13.0
django-cached-field==1.2.15
django-celery==3.2.2
django-ckeditor==5.6.1
django-constance==2.4.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-csvimport==2.11
django-custom-field==2.9
django-dajaxice==0.7
django-daterange-filter==1.3.0
django-debug-toolbar==1.2.2
django-extensions==1.4.6
django-extra-views==0.11.0
django-filter==2.1.0
django-floppy-gumby==0.1
django-floppyforms==1.7.0
django-grappelli==2.12.2
django-haystack==2.8.1
django-heroku-redisify==0.2.1
django-impersonate==1.4
django-jet==1.0.8
django-js-asset==1.2.2
django-localflavor==2.1
django-mass-edit==3.2.0
django-messages==0.5.3
django-model-utils==3.1.2
django-money==0.14.3
django-oscar==1.6.4
django-oscar-accounts==1.0.0
django-oscar-paypal==1.0.0
django-pagination==1.0.7
django-phonenumber-field==2.0.1
django-positions==0.6.0
django-redis-cache==1.8.1
django-report-builder==6.2.2
django-report-scaffold==0.1.4
django-report-utils==0.3.15
django-responsive-dashboard==1.3.3
django-reversion==3.0.3
django-simple-import==2.0.2
django-star-ratings==0.7.0
django-tables2==1.21.2
django-thumbs==0.4
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangocms-admin-style==1.2.9
djangorestframework==3.9.0
djangorestframework-bulk==0.2.1
docutils==0.12
ebaysdk==2.1.5
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.1
feedparser==5.2.1
filters==1.3.2
filters-django==1.0.5
flower==0.9.3
future==0.17.1
gevent==1.3.4
gmane==0.1.dev25
gnureadline==6.3.8
greenlet==0.4.13
gunicorn==19.9.0
hiredis==0.1.5
html2text==2016.9.19
html5lib==1.0.1
httpagentparser==1.8.2
idna==2.7
ipdb==0.8
ipython==2.3.1
isort==4.3.19
jdcal==1.4
Jinja2==2.8.1
jsonfield2==3.0.2
kiwisolver==1.0.1
kombu==3.0.37
lazy-object-proxy==1.4.1
ldap-groups==4.2.2
ldap3==2.5.2
libsass==0.16.0
lxml==4.2.3
Mako==1.0.4
Markdown==3.0.1
MarkupSafe==0.23
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
mysqlclient==1.3.13
networkx==2.2
nltk==3.4
num2words==0.5.6
numpy==1.16.2
ofxparse==0.16
olefile==0.46
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
openpyxl==2.6.1
pandas==0.24.2
parsedatetime==2.4
passlib==1.6.5
pbr==4.3.0
phonenumbers==8.10.0
Pillow==5.4.1
pip-tools==0.3.5
psutil==5.4.8
psycopg2==2.7.1
psycopg2-binary==2.7.6.1
purl==1.4
py-moneyed==0.8.0
pyasn1==0.4.5
pycountry==18.12.8
pycparser==2.19
pylint==2.3.1
PyMySQL==0.9.2
pyparsing==2.3.1
PyPDF2==1.26.0
pysolr==3.8.1
python-dateutil==2.7.5
pytz==2016.7
raven==6.10.0
redis==2.10.6
regex==2019.3.12
reportlab==3.5.12
requests==2.20.0
rosetta==0.3
scipy==1.2.1
selenium==3.141.0
singledispatch==3.4.0.3
six==1.11.0
social==0.1.dev3
sorl-thumbnail==12.4.1
soupsieve==1.6
splinter==0.10.0
sqlparse==0.2.4
stevedore==1.29.0
suds-py3==1.3.3.0
swapper==1.1.0
text-unidecode==1.2
tornado==5.1.1
typed-ast==1.3.5
Unidecode==1.0.23
urllib3==1.24.1
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
webencodings==0.5.1
Werkzeug==0.9.6
Willow==1.1
wrapt==1.11.1
xlwt==1.3.0
